updated about page formating and WIP portfolio
The about page is more responsive and the format has been fine tuned. The portfolio has a completely new design, but still needs work to be more responsive. Donna suggest more color in my website and I need to make sure it is clear the bottom image nav is are links for the corresponding project.